Unity: create embedded package
