--- Unity/Assets/Native/Editor/FragImporter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f976cc32e193649f8a97e05e8d727073
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
