--- Unity/Assets/Native/ShardsController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ecbbc73dc2fdf4b9b8855d2cfeee08ae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
