--- Unity/Assets/Native/test.frag
@@ -1,1 +1,0 @@
-{"ipfsHash": "0x0"}
