--- Unity/Assets/Native/test.frag.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: de77dc5d8afbb46cdb0e026737dee58f
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: f976cc32e193649f8a97e05e8d727073, type: 3}
