--- Unity/Assets/Plugins/.gitignore
@@ -1,2 +1,0 @@
-*.dll
-*.dll.meta
