+++ Unity/Packages/com.fragcolor.shards/CHANGELOG.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
-folderAsset: yes
-DefaultImporter:
+guid: 323867b21f1b1614da15d1279f17583b
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
