+++ Unity/Packages/com.fragcolor.shards/Editor/Fragcolor.Shards.UnityEditor.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "Fragcolor.Shards.UnityEditor",
+    "rootNamespace": "",
+    "references": [
+        "Fragcolor.Shards.UnityEngine"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
