+++ Unity/Packages/com.fragcolor.shards/Editor/Fragcolor.Shards.UnityEditor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 577ba674e31abc04098f343138ce199a
-folderAsset: yes
-DefaultImporter:
+guid: 9fe78b10115507b44afb954e991e9e8a
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
