+++ Unity/Packages/com.fragcolor.shards/LICENSE.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
-folderAsset: yes
-DefaultImporter:
+guid: beb85d8eb52ca1749af843fb0b54b552
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
