+++ Unity/Packages/com.fragcolor.shards/README.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
-folderAsset: yes
-DefaultImporter:
+guid: 98ba311e7c1e5d6499e025a17e1a87e9
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
