+++ Unity/Packages/com.fragcolor.shards/Runtime/Fragcolor.Shards.UnityEngine.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "Fragcolor.Shards.UnityEngine",
+    "rootNamespace": "",
+    "references": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": true,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "Fragcolor.Shards.Common.dll",
+        "System.Runtime.CompilerServices.Unsafe.dll"
+    ],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
