+++ Unity/Packages/com.fragcolor.shards/Runtime/Fragcolor.Shards.UnityEngine.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
-folderAsset: yes
-DefaultImporter:
+guid: 69a256cffb9d0b64a987a4af1829b89c
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
