+++ Unity/Packages/com.fragcolor.shards/Runtime/Plugins.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
+guid: f6e571998de1dd64c92687d3c23cbacf
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
