+++ Unity/Packages/com.fragcolor.shards/Runtime/Plugins/README.md.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fab486c8138c3be4cbe1d57d3a17eaf1
+guid: 62479d668de8e60488817a4478aea339
 TextScriptImporter:
   externalObjects: {}
   userData:
