+++ Unity/Packages/com.fragcolor.shards/Runtime/ShardsController.cs.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 81b4ff339d485814cb44df0b0a432eb8
+guid: 4c960bd5b547b4b448f65d29d910aa1d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: -50
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
