+++ Unity/Packages/com.fragcolor.shards/Runtime/ShardsExtensions.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5bd4bb6c5a67e4f9f9fba1d51d9c89da
+guid: d7e7c671bb3d22745b0862247e8cd66e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
