+++ Unity/Packages/com.fragcolor.shards/Third Party Notices.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 6050634c4c3358a49b22e246cfadd925
-folderAsset: yes
-DefaultImporter:
+guid: 0baebb349cfb90545bc5454e71c77bd2
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
