+++ Unity/Packages/com.fragcolor.shards/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "com.fragcolor.shards",
+    "displayName": "Shards",
+    "version": "0.1.1",
+    "unity": "2020.3",
+    "description": "",    
+    "author": {
+        "name": "Fragcolor",
+        "email": "contact@fragcolor.com",
+        "url": "https://www.fragcolor.com"
+    },
+    "repository": {
+        "url": "https://www.github.com/fragcolor-xyz/sharp-shards.git",
+        "type": "git"
+    }
+}
