+++ Unity/Packages/com.fragcolor.shards/package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 42479972d453436419c83002f6825dae
-folderAsset: yes
-DefaultImporter:
+guid: 030e93c53a71976459aaaef725681c81
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
