+++ Unity/Packages/manifest.json
@@ -1,14 +1,13 @@
 {
   "dependencies": {
     "com.unity.collab-proxy": "1.15.9",
-    "com.unity.ide.rider": "3.0.10",
+    "com.unity.ide.rider": "2.0.7",
     "com.unity.ide.visualstudio": "2.0.14",
     "com.unity.ide.vscode": "1.2.4",
     "com.unity.test-framework": "1.1.30",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.4.8",
     "com.unity.ugui": "1.0.0",
-    "com.unity.visualscripting": "1.7.5",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
