--- Unity/ProjectSettings/Packages/com.unity.testtools.codecoverage/Settings.json
@@ -1,5 +1,0 @@
-{
-    "m_Dictionary": {
-        "m_DictionaryValues": []
-    }
-}
